site_menage_conciergerie/requirements.txt
Django==5.1.1
djangorestframework==3.15.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
Pillow==10.4.0 